Experiment: do different validation splits give more accurate models?
Results: No. All the validation splits tried (with train_test_split) give results in the exact opposite order of the leaderboard.

The most accurate results so far come if we train on X and evaluate on either X or X_val. Both result in the 3 models having the same position as on the leaderboard. This feels like bad ML though... I shouldn't be training models on all data and using that to make predictions, should I?